Fix memory leak in lambda captures
Shared pointers were being captured by value in the lambda capture list.
Capturing by reference introduced subtle bugs. However, when capturing
by value, the reference count is incremented, and the matching decrement
never happens.

This caused a memory leak.

By converting the shared pointers to weak pointers, and capturing those
by value, the reference count problem is fixed.

By using weak pointers, it is also possible to check if the pointer has
been freed before accessing, and throw an appropriate exception.

+++ examples/one_layer.cpp
@@ -53,7 +53,7 @@ read_file(const std::string &filename) {
     std::unordered_map<char, int> stoi;
     std::unordered_map<int, char> itos;
 
-    for (int i = 0; i < chars.size(); i++) {
+    for (size_t i = 0; i < chars.size(); i++) {
         stoi[chars[i]] = i;
         itos[i] = chars[i];
     }
